src/main/java/com/example/voicy_notes/service/UserService.java
package com.example.voicy_notes.service;

import com.example.voicy_notes.entity.User;
import com.example.voicy_notes.exception.BadRequestException;
import com.example.voicy_notes.repository.UserRepository;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

@Service
public class UserService implements UserDetailsService {
    private final UserRepository userRepository;

    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = userRepository.findByUsername(username);
        if(user == null) {
            throw new UsernameNotFoundException("User not found");
        }
        return user;
    }

    public boolean existsByUsername(String username) {
        return userRepository.existsByUsername(username);
    }

    public boolean existsByEmail(String email) {
        return userRepository.existsByEmail(email);
    }

    public User findUserById(Long id) {
        Optional<User> user = userRepository.findById(id);
        return user.orElse(new User());
    }

    public List<User> allUsers() {
        return userRepository.findAll();
    }

    public List<User> allUsersExceptAuthenticated() {
        List<User> result = new ArrayList<>(userRepository.findAll());
        result.remove(findUserById(((User)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId()));
        return result;
    }

    public User saveUser(User user){
        if(userRepository.existsByEmail(user.getEmail())) {
            throw new BadRequestException("Email" + user.getEmail() + " is taken");
        }
        return userRepository.save(user);
    }

    public User findUserByUsername(String username) {
        return userRepository.findByUsername(username);
    }


}
